requirements.txt
# 13 APR 2018 UPDATED

# Need to install below apt-get packages before
# sudo apt-get install python-pip libxml2 libxml2-dev libxslt1-dev libmysqlclient-dev uwsgi nginx
# sudo apt-get build-dep python-imaging
# follow 2 steps link https://docs.confluent.io/current/installation/installing_cp.html#installation-apt
# sudo apt-get install librdkafka++1 librdkafka-dev librdkafka1 confluent-kafka-2.11

amqp==1.4.6
anyjson==0.3.3
attrs==17.4.0
Automat==0.6.0
autopep8==1.3.4
billiard==3.3.0.20
blist==1.3.6
boto==2.40.0
cassandra-driver==3.12.0
cassandra-pylib==0.0.0
cassandra-snapshotter==1.0.0
celery==3.1.18
chardet==2.3.0
configparser==3.3.0.post2
confluent-kafka==0.9.1.2
constantly==15.1.0
crcmod==1.7
datetime-truncate==1.0.1
decorator==4.0.4
defusedxml==0.4.1
Django==1.8.3
django-allauth==0.23.0
django-bootstrap-form==3.2.1
django-bootstrap3==6.2.0
django-braces==1.8.1
django-cassandra-engine==1.3.0
django-compat==1.0.11
django-contrib-comments==1.8.0
django-cors-headers==1.1.0
django-crontab==0.6.0
django-debug-toolbar==1.3.2
django-easy-pdf==0.1.0
django-extensions==1.5.5
django-guardian==1.3
django-hijack==2.1.0
django-hijack-admin==2.1.1
django-markdown-deux==1.0.5
django-organizations==0.5.3
django-pandas==0.4.5
django-recaptcha==1.0.4
django-registration-redux==1.1
django-rest-auth==0.6.0
django-rest-framework-docs==0.1.7
django-rest-swagger==0.3.3
django-screamshot==0.8.1
django-silk==2.0.0
django-tagging==0.4.6
django-timedeltafield==0.7.10
djangorestframework==3.2.2
djangotoolbox==1.6.2
docutils==0.12
Dozer==0.4
drf-nested-routers==0.9.0
ecdsa==0.13
enum34==1.1.6
et-xmlfile==1.0.1
Fabric==1.11.1
funcsigs==1.0.2
futures==3.0.3
geographiclib==1.49
geopy==1.13.0
google-compute-engine==2.7.5
gprof2dot==2016.10.13
gyp==0.1
html5lib==0.9999999
httplib2==0.9.2
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
ipython==4.0.0
ipython-genutils==0.1.0
isoweek==1.3.0
jdcal==1.3
Jinja2==2.7.3
jsonfield==1.0.3
jsonpickle==0.9.2
kombu==3.0.30
librabbitmq==1.6.1
lockfile==0.10.2
lxml==3.4.0
Mako==1.0.1
markdown2==2.3.1
MarkupSafe==0.23
mock==2.0.0
MySQL-python==1.2.5
nose==1.3.7
numpy==1.11.3
oauthlib==1.0.3
openpyxl==2.4.1
pandas==0.19.2
paramiko==1.17.0
path.py==8.1.2
pathlib2==2.3.0
pbr==4.0.3
pexpect==4.0.1
pickleshare==0.5
Pillow==2.6.1
poster==0.8.1
postmarker==0.11.3
psutil==3.2.2
ptyprocess==0.5
pycodestyle==2.3.1
pycrypto==2.6.1
Pygments==2.0.1
pymodbus==1.2.0
PyPDF2==1.25.1
pyserial==3.0
python-daemon==2.0.5
python-dateutil==2.6.0
python-magic==0.4.11
python-memcached==1.54
python-openid==2.2.5
pytz==2016.10
PyYAML==3.11
redis==2.10.3
reportlab==3.2.0
requests==2.7.0
requests-oauthlib==0.5.0
roman==2.0.0
s3cmd==1.6.1
scandir==1.7
scikit-learn==0.18.1
scipy==0.18.0
simplegeneric==0.8.1
simplejson==3.8.2
six==1.10.0
SOAPpy==0.12.22
Sphinx==1.2.3
sqlparse==0.1.16
tabulate==0.7.5
traitlets==4.0.0
twilio==5.4.0
Twisted==15.5.0
Unirest==1.1.7
urllib3==1.13.1
uwsgitop==0.9
virtualenv==13.1.2
WebOb==1.4.1
wstools==0.4.3
xhtml2pdf==0.0.6
xlrd==1.0.0
XlsxWriter==0.9.6
xlwt==1.1.2
zope.interface==4.1.3